requirements.txt
prompt_toolkit>=3.0.43
rich>=13.7.0
openai>=1.0.0
pygments>=2.17.0
python-docx>=1.1.0
fpdf2>=2.7.6
markdown>=3.5.0
